FIX reopen connection after each migration
The migration may contain PRAGMA changes that are not supposed to
propagate to next migration. And in fact, when using the `recreate`
mode, it's not a "may", it's a "will", given we deactivate foreign keys.

Let's execute each migration on its own connection (by reopening the
connection after each migration), so that the PRAGMAs are properly
flushed back to default.